% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKETworkflow.R
\name{ANA_TEST_final}
\alias{ANA_TEST_final}
\title{ANA_TEST_final}
\usage{
ANA_TEST_final(my_dirs, nPERM = 1e+05)
}
\arguments{
\item{my_dirs}{A list generated by \code{setdirs()}
of directories specific to a cancer type}

\item{nPERM}{A positive integer specifying
the number of permutations to calculate
    permutation-based individual kernel and omnibus
p-values}
}
\description{
ANA_TEST_final
}
