% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKETworkflow.R
\name{calc_full_DIST}
\alias{calc_full_DIST}
\title{calc_full_DIST}
\usage{
calc_full_DIST(my_dirs, rmWGA = TRUE, min_gene_mut, LAMBDA, ncores = 1)
}
\arguments{
\item{my_dirs}{A list generated by \code{setdirs()}
of directories specific to a cancer type}

\item{rmWGA}{Boolean set to TRUE to exclude WGA samples}

\item{min_gene_mut}{A non-negative integer to filter
genes, based on mutation frequency, to include in distance 
calculations}

\item{LAMBDA}{A positive numeric value or \code{Inf} to
indicate unbalanced or balanced optimal transport distance
calculation, respectively.}

\item{ncores}{A positive integer to specify the number of
threads to decrease computational runtime.}
}
\description{
calc_full_DIST
}
