% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKETworkflow.R
\name{new_ANA_AGG}
\alias{new_ANA_AGG}
\title{new_ANA_AGG}
\usage{
new_ANA_AGG(my_dirs)
}
\arguments{
\item{my_dirs}{A list generated by \code{setdirs()}
of directories specific to a cancer type}
}
\description{
new_ANA_AGG
}
