% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKETworkflow.R
\name{run_full_geneSim}
\alias{run_full_geneSim}
\title{run_full_geneSim}
\usage{
run_full_geneSim(my_dirs, min_gene_mut, rmWGA = TRUE)
}
\arguments{
\item{my_dirs}{A list generated by \code{setdirs()}
of directories specific to a cancer type}

\item{min_gene_mut}{A non-negative integer to filter
genes, based on mutation frequency, to include in distance 
calculations}

\item{rmWGA}{Boolean set to TRUE to exclude WGA samples}
}
\description{
run_full_geneSim
}
