% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKETworkflow.R
\name{setdirs}
\alias{setdirs}
\title{setdirs}
\usage{
setdirs(git_dir, work_dir, dataset, verbose = TRUE)
}
\arguments{
\item{git_dir}{A string containing the full
path to the parent directory containing
the cloned ROKETworkflow repository}

\item{work_dir}{A string containing the full
path to the working directory for downloading,
processing, and analyzing real data}

\item{dataset}{A string for the abbreviated
cancer type such as BRCA, LGG, STAD, etc.}

\item{verbose}{Boolean set to TRUE to display
verbose output}
}
\description{
setdirs
}
