% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKETworkflow.R
\name{new_ANA_AGG}
\alias{new_ANA_AGG}
\title{new_ANA_AGG}
\usage{
new_ANA_AGG(git_dir, work_dir)
}
\arguments{
\item{git_dir}{A string containing the full
path to the parent directory containing
the cloned ROKETworkflow repository}

\item{work_dir}{A string containing the full
path to the working directory for downloading,
processing, and analyzing real data}
}
\description{
new_ANA_AGG
}
